requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-cors-headers>=4.4
psycopg2-binary>=2.9
django-environ>=0.11
boto3>=1.34
django-storages>=1.14
channels>=4.0
redis>=5.0
uvicorn[standard]>=0.30
gunicorn>=22.0
coverage>=7.6
